Addressed Pavan's review comments 1. Moved the parsing tool and associated UTs to a new folder pkg/tools 2. Fixed the func docs 3. unexported all types from the parsing tool 4. Fixed variable names 5. Printed the entire error stack 6. Fixed spelling issues
